Fix compiler warning C4242: 'argument' : conversion from 'INT_PTR' to 'DWORD', possible loss of data
git-svn-id: svn://svn.code.sf.net/p/shareaza/code/trunk@7873 3231ed4b-2112-0410-84e0-8c3944a40730

+++ shareaza/EDClient.cpp
@@ -1624,7 +1624,7 @@ BOOL CEDClient::OnAskSharedDirs(CEDPacket* /*pPacket*/)
 
 			if ( CEDPacket* pReply = CEDPacket::New( ED2K_C2C_ASKSHAREDDIRSANSWER ) )
 			{
-				pReply->WriteLongLE( oFolderPath.GetCount() );
+				pReply->WriteLongLE( static_cast< DWORD>( oFolderPath.GetCount() ) );
 
 				for ( POSITION pos = oFolderPath.GetHeadPosition(); pos ; )
 				{
